% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skyline_utils.R
\name{getChangeTimes}
\alias{getChangeTimes}
\title{Case 1:
        skyline   = m x n matrix
        times     = m x n matrix
        gridtimes = n-dim vector}
\usage{
getChangeTimes(trees, groupsizes, eventtypes = "coalescent")
}
\description{
Case 1a:
            times[i,1] < gridtimes[1] for all i     -> Add column of Inf to the right of times
        Case 1b:
            times[i,1] >= gridtimes[1] for some i
            times[i,n] >  gridtimes[n] for all i    -> Add column of -Inf to the left of times
        Case 1c:
            times[i,1] >= gridtimes[1] for some i   -> Add column of -Inf to the left of times
            times[i,n] <= gridtimes[n] for some i   -> Change last column of times to Inf
}
\details{
Case 2:
        skyline   = m x n matrix
        times     = m x n-1 matrix
        gridtimes = n-dim vector

times[i,1] > gridtimes[1] for all i
        times[i,n-1] < gridtimes[n] for all i      -> Add column of -Inf to the left and Inf to the right of times

Anything else:
        Error!
}
