% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Skyline_Utilities.R
\name{getSampleTraces}
\alias{getSampleTraces}
\title{Function used for testing}
\usage{
getSampleTraces(true_t, true_N, nsamples = 10, dimension = 5,
  noise_sd = 5, plot = FALSE)
}
\description{
Get sample traces given true time and N vectors, draw <nsamples> piecewise constant
approximations of dimension <dimension> with Gaussian noise with standard deviation
<noise_sd> added.
}
