% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Skyline_Utilities.R
\name{reconstructBackwardSkylineSlow}
\alias{reconstructBackwardSkylineSlow}
\title{Reconstructs a skyline that is measured backwards-in-time (t = 0 is the present)
THIS IS THE SLOW AND CAREFUL FUNCTION THAT ONLY USES PRIMITIVES AND NO VECTORIZATION!}
\usage{
reconstructBackwardSkylineSlow(skyline, times, gridtimes)
}
\arguments{
\item{skyline}{: Skyline estimates in a matrix or data.frame form (rows = MCMC states, cols = skyline estimates during intervals)}

\item{times}{: Times associated with the end-times of each skyline interval (rows = MCMC states, cols = interval start-times)
Must be the same dimension as skyline.}

\item{gridtimes}{: The regular times to reconstruct the marginal posterior on.}

\item{maxtime}{: Maximum time of the skyline function. If this is a scalar it is assumed that the same maxtime is used for every row in skyline}
}
\description{
For state i and interval j, skyline[i,j] is the estimate between times[i,j-1] and times[i,j].
Thus,
}
\details{
skyline   = m x n matrix
    times     = m x n matrix
    gridtimes = n-dim vector

times[i] < times[i+1] (sorted)
    times[i-1] < times[i] (sorted)
    times[i,j] != 0

skyline[i,1] <=> t < times[i,1]                  j == 1 (implicitly set t_0 to 0)
    skyline[i,j] <=> times[i,j-1] < t <= times[i,j]  j > 1
    skyline[i,j] <=> t > times[i,j]                  j == ncol(skyline)
}
