% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skyline_lk.R
\name{skyline_lk}
\alias{skyline_lk}
\title{Generalized Skyline Plot log-likelihood}
\usage{
skyline_lk(tree, popSizes, segmentSizes = NULL, full = TRUE)
}
\description{
Generalized Skyline Plot log-likelihood
}
