% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skyline_lk.R
\name{esp_lk}
\alias{esp_lk}
\title{Epoch Sampling Skyline Plot log-likelihood}
\usage{
esp_lk(tree, popSizes, segmentSizes = NULL, sampIntensities,
  epochSizes = NULL, full = TRUE)
}
\description{
Epoch Sampling Skyline Plot log-likelihood
}
